new tools for tinified pics

+++ tools/tool-box.py
@@ -9,6 +9,7 @@ parser = argparse.ArgumentParser()
 parser.add_argument("-i", "--input", required=True, help="input dir")
 parser.add_argument("-o", "--output", help="out")
 parser.add_argument("-d", "--delete", action="store_true", default=False, help="remove bam and depth files")
+parser.add_argument("--to_cram", action="store_true", default=False, help="transfer bam file to cram")
 parser.add_argument("-lv", "--link_vcf", action="store_true", default=False, help="link gvcf file")
 parser.add_argument("-r", "--recal_bam", action="store_true", default=False, help="remove recal_bam files")
 args = parser.parse_args()
@@ -35,7 +36,10 @@ def scan_bam(path):
     for entry in os.scandir(path):
         if entry.is_dir() and (entry.name == 'bam' or entry.name == 'depth'):
             print (entry.path)
-            removeDir(entry.path)
+            if entry.name == 'bam' and args.to_cram:
+                toCram(entry.path)
+            if args.delete:
+                removeDir(entry.path)
         elif args.recal_bam and entry.is_dir() and entry.name == 'recal_bam':
             print (entry.path)
             removeDir(entry.path)
@@ -63,7 +67,24 @@ def vcf_file(path):
     else:
         sys.exit('link file error')
 
-if args.delete:
+def toCram(path):
+    if os.path.isdir(path):
+        cramDir = path.replace('bam', 'cram')
+        if not os.path.exists(cramDir):
+            try:
+                os.makedirs(cramDir)
+            except:
+                sys.exit('cram dir created failed')
+        for entry in os.scandir(path):
+            if entry.is_file() and '.bam' in entry.name:
+                cramFile = cramDir + '/' + entry.name.replace('.bam', '.cram')
+                cramCmd = ['samtools', 'view', '-@', '20', '-T', '/opt/seqtools/gatk/ucsc.hg19.fasta', '-C', '-o', cramFile, entry.path]
+                step = Popen(cramCmd)
+                step.wait()
+    else:
+        sys.exit('bam dir error')
+
+if args.delete or args.to_cram:
     scan_bam(inputDir)
     print ('remove process done!')
